refactor: merge Main/types.ts into algorithms/types.ts
