+++ packages/web/src/components/Main/MutationTooltip.tsx
@@ -2,8 +2,7 @@ import React from 'react'
 
 import { Popover, PopoverBody } from 'reactstrap'
 
-import type { MutationElementWithId } from 'src/components/Main/types'
-import type { AnalysisResult } from 'src/algorithms/types'
+import type { AnalysisResult, MutationElementWithId } from 'src/algorithms/types'
 
 export interface MutationTooltipProps {
   sequence: AnalysisResult
