--- packages/web/src/components/Main/types.ts
@@ -1,20 +1,0 @@
-export interface MutationElement {
-  seqName: string
-  positionZeroBased: string
-  allele: string
-}
-
-export interface MutationElementWithId extends MutationElement {
-  id: string
-}
-
-export interface MissingElement {
-  seqName: string
-  character: string
-  begin: number
-  end: number
-}
-
-export interface MissingElementWithId extends MissingElement {
-  id: string
-}
